Refactor ghost story state into hook
